Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
requests
qdrant-client
pypdf
python-docx
sqlalchemy